Update README.md
Today I started on the addition and subtraction part of this fraction part.  One of the things I wanted students to realize is that the addition problem 2 + -3 is the same as the subtraction problem 2 - 3.  Similarly the subtraction problem -2 - 3 is the same as the addition problem -2 + (-3).  So, I am working on making sure the student can figure out how to do these problems step by step.  Hoping to finish this tomorrow.